Added Read me file and some other changes
