+++ src/test/java/com/bh/assign/runner/RunnerTest.java
@@ -4,8 +4,7 @@ import io.cucumber.testng.CucumberOptions;
 
 @CucumberOptions(
         plugin = {"pretty", "html:target/cucumber-report.html","json:target/cucumber/Destination/cucumber.json"},
-        monochrome = true
-        ,
+        monochrome = true,
 //        tags="@BillingDemo",
         features = {"src/test/resources"},glue={"com.bh.assign.stepdef"}
 )
