+++ src/test/resources/feature.api/demoabcbilling.feature
@@ -3,7 +3,7 @@ Feature:Demo abc billing rest api test validation
 
 #######################################
 
-	@Get_Access_token_Post
+	@DemoBilling_Get_Access_token_Post
 	Scenario Outline: get billing access token be sending API requests of type Post to grant access to server
 		Given user hit the "<uri>" and "<jsonFilepath>" apiEndpoint with header details and make Post call
 #	|header|contentType|application/json|
@@ -17,20 +17,20 @@ Feature:Demo abc billing rest api test validation
 			| uri |jsonFilepath|expectedStatusCode |expectedKey |expectedField1 |expectedValue1 |expectedField2 |expectedValue2 |
 			|grantAccessUrl|src/test/resources/util/Jsonfile/billinGrantaccess.json|200 |access_token |token_type | bearer |expires_in |899  |
 
-	@Scenario_postAPI
-	Scenario: Send API requests of  type Post to server
+	@DemoBilling_postRetrieveAPI
+	Scenario: Send API requests of  type Post to server to retrieve all client profile
 #	Given user hit the uri apiEndpoint
 		When User provide header details and make Post call for grant access token
 		When User provide header details and make get call to retrieve all with access provided
 
-	@Scenario_get_PaymentMethod
-	Scenario: Send API requests of  type get for Payment method to server
+	@DemoBilling_get_PaymentMethod
+	Scenario: Send API requests of  type get for Payment method to server for payment method
 #	Given user hit the uri apiEndpoint
 		When User provide header details and make Post call for grant access token
 		When User provide header details and make get call to Payment method with access provided
 
-	@Scenario_post_PaymentMethod
-	Scenario: Send API requests of  type Post for Payment method to server
+	@DemoBilling_post_PaymentMethod
+	Scenario: Send API requests of  type Post for Payment method to server for payment method
 #	Given user hit the uri apiEndpoint
 		When User provide header details and make Post call for grant access token
 		When User provide header details and make Post call to Payment method with access provided
